Initial implementation of per-counter timed backgrounds
Use similiar logic to the voice messages for changing the clock backgrounds.

We will look at the bg/timer[0-9]/ directories for images;
default.* will be used as the default BG when this timer is running,
otherwise pictures in the form of 001.* will be displayed when a countdown
timer passes that second mark, like with the sound assets.